third_party/WebKit/LayoutTests/usb/usb.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/fake-devices.js"></script>
<script src="resources/usb-helpers.js"></script>
<script src="resources/webusb-test.js"></script>
<script>
'use strict';

// Returns a promise that is resolved when the next USBConnectionEvent of the
// given type is received.
function connectionEventPromise(eventType) {
  return new Promise(resolve => {
    let eventHandler = e => {
      assert_true(e instanceof USBConnectionEvent);
      navigator.usb.removeEventListener(eventType, eventHandler);
      resolve(e.device);
    };
    navigator.usb.addEventListener(eventType, eventHandler);
  });
}

usb_test(() => {
  navigator.usb.test.addFakeDevice(fakeDeviceInit);

  return navigator.usb.getDevices().then(devices => {
    assert_equals(devices.length, 1);
    assertDeviceInfoEquals(devices[0], fakeDeviceInit);
  });
}, 'getDevices returns devices exposed by the DeviceManager service');

usb_test(() => {
  navigator.usb.test.addFakeDevice(fakeDeviceInit);

  return navigator.usb.getDevices().then(devicesFirstTime => {
    assert_equals(devicesFirstTime.length, 1);
    return navigator.usb.getDevices().then(devicesSecondTime => {
      assert_array_equals(devicesSecondTime, devicesFirstTime);
    });
  });
}, 'getDevices returns the same objects for each USB device');

usb_test(() => {
  return navigator.usb.requestDevice({ filters: [] })
    .then(device => {
      assert_unreachable('requestDevice should reject without a user gesture');
    })
    .catch(error => {
      assert_equals(error.code, DOMException.SECURITY_ERR);
    });
}, 'requestDevice rejects when called without a user gesture');

usb_test(() => {
  return callWithKeyDown(() => navigator.usb.requestDevice({ filters: [] })
    .then(device => {
      assert_unreachable('requestDevice should reject when no device selected');
    })
    .catch(error => {
      assert_equals(error.code, DOMException.NOT_FOUND_ERR);
    })
  );
}, 'requestDevice rejects when no device is chosen');

usb_test(() => {
  let guid = navigator.usb.test.addFakeDevice(fakeDeviceInit);
  navigator.usb.test.chosenDevice = guid;

  return callWithKeyDown(() => navigator.usb.requestDevice({ filters: [] })
    .then(device => {
      assertDeviceInfoEquals(device, fakeDeviceInit);
    })
  );
}, 'requestDevice returns the device chosen by the user');

usb_test(() => {
  let guid = navigator.usb.test.addFakeDevice(fakeDeviceInit);
  navigator.usb.test.chosenDevice = guid;

  return callWithKeyDown(() => navigator.usb.requestDevice({ filters: [] })
      .then(device => navigator.usb.getDevices().then(devices => {
        assert_equals(devices.length, 1);
        assert_equals(devices[0], device);
      }))
  );
}, 'getDevices returns the same object as requestDevice');

usb_test(() => {
  const expectedFilters = [
    { vendorId: 1234, classCode: 0xFF, serialNumber: "123ABC" },
    { vendorId: 5678, productId: 0xF00F }
  ];

  return callWithKeyDown(() => navigator.usb.requestDevice({ filters: expectedFilters })
    .then(device => {
      assert_unreachable('requestDevice should reject because no device selected');
    })
    .catch(error => {
      assert_equals(error.code, DOMException.NOT_FOUND_ERR);
      let actualFilters = navigator.usb.test.lastFilters;
      assert_equals(actualFilters.length, expectedFilters.length);
      for (var i = 0; i < actualFilters.length; ++i)
        assert_object_equals(actualFilters[i], expectedFilters[i]);
    })
  );
}, 'filters are sent correctly');

usb_test(() => {
  let promise = connectionEventPromise('connect');
  navigator.usb.test.addFakeDevice(fakeDeviceInit);

  return promise.then(device => {
    assertDeviceInfoEquals(device, fakeDeviceInit);
    return device.open().then(() => device.close());
  });
}, 'onconnect event is trigged by adding a device');

usb_test(usb => {
  let promise = connectionEventPromise('connect');
  navigator.usb.test.addFakeDevice(fakeDeviceInit);

  return promise
      .then(device => navigator.usb.getDevices().then(devices => {
        assert_equals(devices.length, 1);
        assert_equals(devices[0], device);
      }));
}, 'getDevices returns the same object as sent in the onconnect event');

usb_test(usb => {
  let deviceAdded = connectionEventPromise('connect');
  let guid = navigator.usb.test.addFakeDevice(fakeDeviceInit);

  return deviceAdded.then(device => {
    let deviceRemoved = connectionEventPromise('disconnect');
    navigator.usb.test.removeFakeDevice(guid);
    return deviceRemoved.then(removedDevice => {
      assertDeviceInfoEquals(removedDevice, fakeDeviceInit);
      assert_equals(removedDevice, device);
      return removedDevice.open().then(() => {
        assert_unreachable('should not be able to open a disconnected device');
      }, error => {
        assert_equals(error.code, DOMException.NOT_FOUND_ERR);
      });
    });
  });
}, 'ondisconnect event is triggered by removing a device');
</script>
